requirements.txt
matplotlib==3.8.4
networkx==3.3
node2vec==0.5.0
numpy==2.1.1
optuna==2.10.1
pandas==2.2.2
Pillow==10.4.0
python_louvain==0.16
scikit_learn==1.4.2
torch==2.2.2
torch_geometric==2.5.3
tqdm==4.66.2